RuneRunner/cv-module/requirements.txt
flask
torch
torchvision
pillow
opencv-python-headless
python-dotenv
requests